+++ CHANGELOG.md
@@ -3,3 +3,5 @@
 * Unreleased
     - New features
         - Multiple photo support. #107
+    - Development improvements
+        - CONFIG.SKIP_CONFIRM_REPORT flag to skip confirmation screen